feat: 供应商评价页面标杆优化及样本数据
- 重写 IqcSuppliersPage.vue，采用标准列表表格模式（对标 IqcReceiptsPage）
- 新增统计栏（A/B/C/D 级供应商数量统计）
- 支持关键词搜索和评级筛选
- 新增评分编辑抽屉（含四维度评分 JSON）
- 清理 IqcTab.vue 中重复的供应商评分 tab
- 修复前端过滤条件（status → isActive）
- 修复评估日期显示格式（仅显示日期，不显示时间）

数据库增强:
- 新增 migration_s3_supplier_scores_enhance.sql 补齐缺失字段
- 新增 seed_supplier_scores.sql 含 20 家供应商评分数据（维度评分 + 评估意见）
- 新增 seed_iqc_receipts_enhanced.sql 补充 15 条 IQC 检验记录

后端修复:
- 修复 SupplierScore 模型映射（移除 NotMapped 标记）
- 修复 score 字段类型（int → decimal）

+++ backend/Controllers/M02_5/DynamicParamsController.cs
@@ -127,7 +127,7 @@ public class DynamicParamsController : ControllerBase
             TargetValue = dto.TargetValue,
             Usl = dto.Usl,
             Lsl = dto.Lsl,
-            Precision = dto.Precision,
+            Precision = (int)dto.Precision,
             AiStrategy = dto.AiStrategy,
             SortOrder = dto.SortOrder,
             CreatedBy = 1, // TODO: 从 JWT 获取
@@ -156,7 +156,7 @@ public class DynamicParamsController : ControllerBase
         entity.TargetValue = dto.TargetValue;
         entity.Usl = dto.Usl;
         entity.Lsl = dto.Lsl;
-        entity.Precision = dto.Precision;
+        entity.Precision = (int)dto.Precision;
         entity.AiStrategy = dto.AiStrategy;
         entity.SortOrder = dto.SortOrder;
         entity.IsActive = dto.IsActive;
